Sanitize URL well
This is a security update and I highly urge you to update your code immediately to stay secure.

This fixes the following security issues
 - XSS

These issue may have a "High" impact as they may compromise the confidentiality, availability, and integrity of servers to the victim.

This update also includes a new method, `mrk.sanitizeURL(url)`, which should be used for all `src` and `href` attributes. You must use this **together with** the existing HTML sanitizer.

This pull request has not been tested.